Use a list of binary trees to store the paths calculated already
Arrange the configuration with non-zero paths in the form of a tree for
each node in the problem.
Each node has a tree that represents the numberOfPaths from it the
destination, given the nodes already visited.
Not storing the results for cases when no path was found because most of
those cases will be found soon so it was only making the tree huge,
which increased the lookup time.